client: add player total score to dex & remove individual planet scores
